starting to make the handleSave function variable on PipelinEditPage
