+++ web-app/src/utils/types.tsx
@@ -55,6 +55,7 @@ export interface AppPipelineInfoType {
 }
 
 export interface AppPipelineType {
+	id: number,
 	name: string;
 	action: AppAREAType;
 	reactions: Array<AppAREAType>;
